requirements.txt
flask>=3.0.0
anthropic>=0.39.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
